refactor(exit-capture): move file convention + reader to mvm-core (dep direction)
WORKLOAD_EXIT_FILE, exit_file_path, read_captured now live in
mvm-core::exit_capture so mvm-backend can read the captured exit
code without depending on mvm-vm-host (which depends on mvm-backend —
a cycle). capture_once stays in mvm-vm-host and re-exports the mvm-core
items for backwards compatibility. Plan 152 WS-A Task 3.0.
