+++ crates/mvm-vm-host/src/exit_capture.rs
@@ -1,26 +1,19 @@
-//! Backend-agnostic workload exit-code capture (Plan 152 WS-A).
+//! Supervisor-side workload exit-code capture (Plan 152 WS-A).
 //!
-//! The guest `/init` writes a 4-byte little-endian `i32` to the control
-//! vsock port before `poweroff -f`. The supervisor binds a host
-//! `UnixListener` at the control socket (libkrun `add_vsock_port2(
-//! listen=false)`), accepts one connection, reads the code, and persists
-//! it to `<vm_state_dir>/workload.exit`. The backend reads that file
-//! after the VM stops. WS-B's Vz supervisor reuses this module.
+//! Binds nothing itself — the caller passes a bound `UnixListener` for
+//! the control vsock port. Reads the guest's 4-byte LE i32 and persists
+//! it via `mvm_core::exit_capture::exit_file_path`. The file convention +
+//! reader live in `mvm-core` so `mvm-backend` can read without depending
+//! on this (supervisor) crate.
 
 use std::io::Read;
 use std::os::unix::net::UnixListener;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
-/// File name under `vm_state_dir` holding the captured exit code (decimal).
-pub const WORKLOAD_EXIT_FILE: &str = "workload.exit";
-
-pub fn exit_file_path(vm_state_dir: &Path) -> PathBuf {
-    vm_state_dir.join(WORKLOAD_EXIT_FILE)
-}
+pub use mvm_core::exit_capture::{WORKLOAD_EXIT_FILE, exit_file_path, read_captured};
 
 /// Block on `listener` for one guest connection, read the 4-byte LE i32,
 /// and persist it to `<vm_state_dir>/workload.exit`. Returns the code.
-/// Best-effort: any error leaves no file (read as "unknown" downstream).
 pub fn capture_once(listener: &UnixListener, vm_state_dir: &Path) -> std::io::Result<i32> {
     let (mut stream, _addr) = listener.accept()?;
     let mut buf = [0u8; 4];
@@ -28,13 +21,6 @@ pub fn capture_once(listener: &UnixListener, vm_state_dir: &Path) -> std::io::Re
     let code = i32::from_le_bytes(buf);
     std::fs::write(exit_file_path(vm_state_dir), code.to_string())?;
     Ok(code)
-}
-
-/// Read a previously-captured exit code, if present.
-pub fn read_captured(vm_state_dir: &Path) -> Option<i32> {
-    std::fs::read_to_string(exit_file_path(vm_state_dir))
-        .ok()
-        .and_then(|s| s.trim().parse().ok())
 }
 
 #[cfg(test)]
@@ -60,11 +46,5 @@ mod tests {
         handle.join().unwrap();
         assert_eq!(code, -7);
         assert_eq!(read_captured(dir.path()), Some(-7));
-    }
-
-    #[test]
-    fn read_captured_is_none_when_absent() {
-        let dir = tempfile::tempdir().unwrap();
-        assert_eq!(read_captured(dir.path()), None);
     }
 }
